+++ tests/test_misc.py
@@ -921,18 +921,25 @@ def test_method_view(app, client):
 
 
 def test_phone_util_override(app, sqlalchemy_datastore):
-    class MyPhoneUtil:
-        def __init__(self, app):
-            pass
+    from flask_security import phone_util
 
+    class MyPhoneUtil(phone_util.PhoneUtil):
         def validate_phone_number(self, input_data):
             return "call-me"
 
         def get_canonical_form(self, input_data):
             return "very-canonical"
 
-    app.security = Security()
-    app.security.init_app(app, sqlalchemy_datastore, phone_util_cls=MyPhoneUtil)
+    app.security = Security(phone_util_cls=MyPhoneUtil)
+    app.security.init_app(app, sqlalchemy_datastore)
+
+    with app.app_context():
+        assert uia_phone_mapper("55") == "very-canonical"
+
+    # try init_app kwargs
+    app.config["SECURITY_BLUEPRINT_NAME"] = "security2"
+    app.security2 = Security()
+    app.security2.init_app(app, sqlalchemy_datastore, phone_util_cls=MyPhoneUtil)
 
     with app.app_context():
         assert uia_phone_mapper("55") == "very-canonical"